+++ upcoming_changelogs/6311.md
@@ -1,3 +1,3 @@
 **CSS-in-JS conversions**
 
-- Converted `EuiTabs` EuiCodeBlock` to Emotion
+- Converted `EuiTabs` to Emotion
